pyproject.toml
[tool.poetry]
name = "punch-chat-server-app"
version = "0.1.0"
description = ""
authors = ["Merrick1307 <holuworley.yz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.14"
pydantic = {extras = ["email"], version = "^2.12.5"}
fastapi = "^0.128.0"
python-multipart = "^0.0.21"
asyncpg = "^0.31.0"
python-dotenv = "^1.2.1"
redis = {version = "^5.2.0", extras = ["hiredis"]}
yoyo-migrations = "^9.0.0"
orjson = "^3.11.5"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
bcrypt = "^5.0.0"
psycopg2-binary = "^2.9.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
